Improve dotadd function
Now `dotadd` shows both untracked and unstaged files (with an empty line between them). Untracked files are previewed with `bat` and unstaged with `git diff`. Note that due to the dotfiles repo being a bare repo, `bat` does not recognize this as a git directory and cannot show the diff.