fix: Remove flaky automatic git repository syncing
Syncing git repos is hard to do automatically. Handling conflicts and
making sure everything is really synced is really tedious and possibly
cannot be done reliably.

Also moved email sending to the top of the sync script. Logical, since
this is the most important operation to get done (e.g. in case of
errors).